refactor: extract ContinueOnFailureFeature from ScenarioFeature
