+++ src/test/Xbehave.Features.Net45/Xbehave.Features.Net45.csproj
@@ -67,6 +67,9 @@
     <Compile Include="..\Xbehave.Features.Net40\BackgroundFeature.cs">
       <Link>BackgroundFeature.cs</Link>
     </Compile>
+    <Compile Include="..\Xbehave.Features.Net40\ContinueOnFailureFeature.cs">
+      <Link>ContinueOnFailureFeature.cs</Link>
+    </Compile>
     <Compile Include="..\Xbehave.Features.Net40\ExampleFeature.cs">
       <Link>ExampleFeature.cs</Link>
     </Compile>
